+++ src/test/java/duke/gfx/SpriteTest.java
@@ -9,7 +9,7 @@ class SpriteTest {
     @Test
     void shouldValidateDimensions() {
         assertThatThrownBy(() ->
-                new Sprite(32, 32, new int[]{0})
+                new Sprite(32, 32, new byte[]{0})
         ).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("Sprite pixels do not match given dimensions");
     }
@@ -18,36 +18,36 @@ class SpriteTest {
     void shouldCreateBlank() {
         Sprite sprite = new Sprite(1, 1);
 
-        assertThat(sprite.getPixel(0, 0)).isEqualTo(0);
+        assertThat(sprite.getPixel(0, 0)).isEqualTo((byte) 0);
     }
 
     @Test
     void shouldClear() {
-        Sprite sprite = new Sprite(1, 1, new int[]{1});
+        Sprite sprite = new Sprite(1, 1, new byte[]{1});
 
         sprite.clear();
 
-        assertThat(sprite.getPixel(0, 0)).isEqualTo(0);
+        assertThat(sprite.getPixel(0, 0)).isEqualTo((byte) 0);
     }
 
     @Test
     void shouldRetrievePixels() {
-        Sprite sprite = new Sprite(1, 1, new int[]{3});
+        Sprite sprite = new Sprite(1, 1, new byte[]{3});
 
         assertThat(sprite.getWidth()).isEqualTo(1);
         assertThat(sprite.getHeight()).isEqualTo(1);
-        assertThat(sprite.getPixel(0, 0)).isEqualTo(3);
+        assertThat(sprite.getPixel(0, 0)).isEqualTo((byte) 3);
     }
 
     @Test
     void shouldOverlay() {
-        Sprite sprite = new Sprite(1, 1, new int[]{1});
-        Sprite overlay = new Sprite(1, 1, new int[]{3});
+        Sprite sprite = new Sprite(1, 1, new byte[]{1});
+        Sprite overlay = new Sprite(1, 1, new byte[]{3});
 
         sprite.draw(overlay, 0, 0);
 
         assertThat(sprite.getWidth()).isEqualTo(1);
         assertThat(sprite.getHeight()).isEqualTo(1);
-        assertThat(sprite.getPixel(0, 0)).isEqualTo(3);
+        assertThat(sprite.getPixel(0, 0)).isEqualTo((byte) 3);
     }
 }
